Change to float field

+++ attractions/models.py
@@ -51,5 +51,5 @@ class Review(models.Model):
     review_text = models.TextField() 
     reviewer_name = models.CharField(max_length=200,blank=True)
     review_title = models.CharField(max_length=200)
-    review_rating = models.PositiveIntegerField(default=0)
+    review_rating = models.FloatField(default=0.0)
     review_date = models.DateTimeField(auto_now_add=True)
